Centralize the overlaps of resource config forms
refs #3044
